Fix re-review issues: separate on-chain from DB writes, remove dead code
Critical fixes:
- Phase 1: separate on-chain tx and DB mark into distinct try blocks.
  On-chain success adds to settled_positions regardless of DB outcome.
  Log message accurately reflects what failed (tx vs DB write).
- Phase 2: separate physicalRedeem tx from DB update. A DB failure after
  successful delivery no longer triggers the failure catch block that
  would incorrectly mark the position as physical_failed.
- Unified _db_update helper with error logging and row-match check,
  used consistently across Phase 2 and OTM updates.
- compute_max_collateral_spent returns (max_collateral, contra_amount)
  to eliminate duplicated delivered_amount arithmetic.
- get_expired_unsettled filters out rows with NULL metadata fields
  (strike_price, is_put, amount) to prevent crashes in identify_itm.
- _store_events raises on silent Supabase data loss (non-empty input,
  empty response) to prevent block pointer from advancing past lost events.
- _update_delivery_events wraps individual writes in try/except so one
  failure doesn't skip remaining delivery events.
- _compute_outcome handles physical_failed: "delivery failed, pending review"
- expiry_price_str uses None instead of "" for missing values

Cleanup:
- Removed dead code: _mark_settled, _mark_physical_delivery (replaced
  by _db_update and _mark_batch_settled), get_controller, CONTROLLER_ABI
  import, underlyingAsset ABI entry
- OnChainOrder.strike_price type: float → int

+++ src/models/order.py
@@ -19,7 +19,7 @@ class OnChainOrder(BaseModel):
     premium: str         # raw USDC premium (6 decimals) as string
     collateral: str      # raw collateral as string
     vault_id: int
-    strike_price: float | None = None
+    strike_price: int | None = None
     expiry: int | None = None
     is_put: bool | None = None
     is_settled: bool = False
